EXAMPLETABLE implemented, with lots of simplemath examples
It was getting awkward to demonstrate simplemath effectively, so I put
lots of examples in a table. The command EXAMPLETABLE handles that.

The examples are not yet referred to in the text, which needs to be
looked at afresh with these in mind.

The EXAMPLETABLE command might want some further work regarding options
etc.