Fixed spacing issue in LBTEXAMPLE
It appeared that a blank line was being output at the end of the
Verbatim environment. Right-stripping the input text didn't make any
difference. ChatGPT told me there is an interaction between Verbatim and
tcolorbox, and I should add `after upper=\vspace{-2pt}` to the
tcolorbox. I did this, but made it -12pt, and it worked.

So it could be fragile, but it's something.